requirements.txt
# NewAppiumTestProject/requirements.txt
pytest==8.3.4
pytest-xdist==3.6.1
allure-pytest==2.13.5
pytest-html==4.1.1
pytest-order==1.3.0
appium-python-client==4.5.1
faker==15.0.0
selenium~=4.29.0
setuptools~=70.0.0
behave~=1.2.6
requests~=2.32.0
assertpy~=1.1